requirements.txt
# RSS News Fetcher Dependencies
feedparser>=6.0.10
chromadb>=0.4.22
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
google-generativeai>=0.3.1
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.3
pandas>=2.1.0
yfinance>=0.2.33
python-dateutil>=2.8.2
numpy>=1.26.0
